Add files via upload
This is a responsive education website designed to provide students with a rich learning experience. The platform offers a user-friendly interface with easy navigation for students, educators, and visitors to explore courses, resources, and educational content.

Features:
Home Page: Introduction to the website with a clean layout showcasing courses, featured content, and announcements.
about: story behind this design.
Services: A detailed list of available courses with descriptions, categories, and filters.
to-do-list : to do list for planning you study routine.

Tech Stack:

HTML5/CSS3: Structure and styling of the website.
JavaScript: Interactive elements and dynamic content.